refactor: extract fatal output routing
